+++ example/reflection_factory_bridge_example.reflection.g.dart
@@ -1,6 +1,6 @@
 //
 // GENERATED CODE - DO NOT MODIFY BY HAND!
-// BUILDER: reflection_factory/1.2.16
+// BUILDER: reflection_factory/1.2.17
 // BUILD COMMAND: dart run build_runner build
 //
 
@@ -35,7 +35,7 @@ class User$reflection extends ClassReflection<User> {
   Version get languageVersion => Version.parse('2.17.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.16');
+  Version get reflectionFactoryVersion => Version.parse('1.2.17');
 
   @override
   User$reflection withObject([User? obj]) => User$reflection(obj);